Updated Hardware
Designed two independent models of casing, clip and velcro. Both models have gone through 2 generations of revisions. The conclusion is that "Thick Clips" are the way to go. This is probably the housing that I will be using for the final product, possibly with the addition of standoffs for the PCB, but we'll see about that.